backend/requirements.txt
# Backend dependencies for Manim Scene Generator
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
websockets==11.0.3
httpx==0.25.0
langchain==0.0.310
openai==1.2.3
numpy==1.26.1
pytest==7.4.3
manim==0.17.3
moviepy==1.0.3
pyngrok==6.0.0
sqlalchemy==2.0.22
alembic==1.12.0